+++ Laura_Fargate/capstone/src/main/java/com/immune/capstone/web/client/model/GasConsumptionSummary.java
@@ -15,7 +15,7 @@ public class GasConsumptionSummary {
 
     private int month;
     private int year;
-    private double consumption;
+    private float consumption;
     private String zone;
 
 }
